rename index.tsx to component names
