+++ src/view/checkout/CheckoutView.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
-import { ProductList } from './components/product-list'
+import { ProductList } from './components/product-list/productList'
 import { UserContext } from '../../shared/providers/UserProvider'
 import { useCart } from 'hooks/useCart'
-import { Button } from 'components/button'
+import { Button } from 'components/button/button'
 import { Product } from 'shared/interfaces/ProductsInterface'
 import { IForm, IInputFieldError, TInputValue } from 'shared/interfaces/form'
 import { useFormRender } from 'hooks/useFormRender'
